rename updated spot the diff file
