+++ Makefile
@@ -2,4 +2,4 @@ cluster-scan:
 	./memory-scanner.sh ${TAG}
 
 cluster-diff:
-	python3 ./spot-the-difference-spicy.py -f ${FIRST_FILE} -s ${SECOND_FILE} -o ${OUTPUT_TAG}
+	python3 ./spot-the-difference.py -f ${FIRST_FILE} -s ${SECOND_FILE} -o ${OUTPUT_TAG}
