--- config/ignore-old.json
@@ -1,292 +1,0 @@
-[
-  {
-    "name": "something",
-    "kind": "something",
-    "namespace": "something"
-  },
-  {
-    "name": "node-exporter",
-    "kind": "securitycontextconstraints.security.openshift.io"
-  },
-  {
-    "name": "openshift-cluster-monitoring-view",
-    "kind": "clusterroles.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "openshift-cluster-monitoring-admin",
-    "kind": "clusterroles.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "openshift-cluster-monitoring-edit",
-    "kind": "clusterroles.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "alertmanager-main",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "prometheus-adapter",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "prometheus-adapter-view",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "prometheus-k8s",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "prometheus-operator",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "resource-metrics:system:auth-delegator",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "telemeter-client",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "telemeter-client-view",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "thanos-querier",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "openshift-state-metrics",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "node-exporter",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "kube-state-metrics",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "grafana",
-    "kind": "clusterrolebindings.rbac.authorization.k8s.io"
-  },
-  {
-    "name": "multus.openshift.io",
-    "kind": "validatingwebhookconfigurations.admissionregistration.k8s.io"
-  },
-  {
-    "name": "clusternetworks.network.openshift.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "name": "egressnetworkpolicies.network.openshift.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "name": "hostsubnets.network.openshift.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "name": "ippools.whereabouts.cni.cncf.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "name": "netnamespaces.network.openshift.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "name": "network-attachment-definitions.k8s.cni.cncf.io",
-    "kind": "customresourcedefinitions.apiextensions.k8s.io"
-  },
-  {
-    "kind": "clusterrolebindings.authorization.openshift.io"
-  },
-  {
-    "kind": "clusterroles.authorization.openshift.io"
-  },
-  {
-    "kind": "clusteroperators.config.openshift.io"
-  },
-  {
-    "kind": "clusterversions.config.openshift.io"
-  },
-  {
-    "kind": "nodes.metrics.k8s.io"
-  },
-  {
-    "kind": "nodes"
-  },
-  {
-    "kind": "events"
-  },
-  {
-    "namespace": "kube-node-lease"
-  },
-  {
-    "namespace": "kube-public"
-  },
-  {
-    "namespace": "kube-system"
-  },
-  {
-    "namespace": "openshift"
-  },
-  {
-    "namespace": "openshift-apiserver"
-  },
-  {
-    "namespace": "openshift-apiserver-operator"
-  },
-  {
-    "namespace": "openshift-authentication"
-  },
-  {
-    "namespace": "openshift-authentication-operator"
-  },
-  {
-    "namespace": "openshift-cloud-credential-operator"
-  },
-  {
-    "namespace": "openshift-cluster-machine-approver"
-  },
-  {
-    "namespace": "openshift-cluster-node-tuning-operator"
-  },
-  {
-    "namespace": "openshift-cluster-samples-operator"
-  },
-  {
-    "namespace": "openshift-cluster-storage-operator"
-  },
-  {
-    "namespace": "openshift-cluster-version"
-  },
-  {
-    "namespace": "openshift-config"
-  },
-  {
-    "namespace": "openshift-config-managed"
-  },
-  {
-    "namespace": "openshift-config-operator"
-  },
-  {
-    "namespace": "openshift-console"
-  },
-  {
-    "namespace": "openshift-console-operator"
-  },
-  {
-    "namespace": "openshift-controller-manager"
-  },
-  {
-    "namespace": "openshift-controller-manager-operator"
-  },
-  {
-    "namespace": "openshift-dns"
-  },
-  {
-    "namespace": "openshift-dns-operator"
-  },
-  {
-    "namespace": "openshift-etcd"
-  },
-  {
-    "namespace": "openshift-etcd-operator"
-  },
-  {
-    "namespace": "openshift-image-registry"
-  },
-  {
-    "namespace": "openshift-infra"
-  },
-  {
-    "namespace": "openshift-ingress"
-  },
-  {
-    "namespace": "openshift-ingress-operator"
-  },
-  {
-    "namespace": "openshift-insights"
-  },
-  {
-    "namespace": "openshift-kni-infra"
-  },
-  {
-    "namespace": "openshift-kube-apiserver"
-  },
-  {
-    "namespace": "openshift-kube-apiserver-operator"
-  },
-  {
-    "namespace": "openshift-kube-controller-manager"
-  },
-  {
-    "namespace": "openshift-kube-controller-manager-operator"
-  },
-  {
-    "namespace": "openshift-kube-scheduler"
-  },
-  {
-    "namespace": "openshift-kube-scheduler-operator"
-  },
-  {
-    "namespace": "openshift-kube-storage-version-migrator"
-  },
-  {
-    "namespace": "openshift-kube-storage-version-migrator-operator"
-  },
-  {
-    "namespace": "openshift-machine-api"
-  },
-  {
-    "namespace": "openshift-machine-config-operator"
-  },
-  {
-    "namespace": "openshift-marketplace"
-  },
-  {
-    "namespace": "openshift-monitoring"
-  },
-  {
-    "namespace": "openshift-multus"
-  },
-  {
-    "namespace": "openshift-network-operator"
-  },
-  {
-    "namespace": "openshift-node"
-  },
-  {
-    "namespace": "openshift-openstack-infra"
-  },
-  {
-    "namespace": "openshift-operator-lifecycle-manager"
-  },
-  {
-    "namespace": "openshift-operators"
-  },
-  {
-    "namespace": "openshift-ovirt-infra"
-  },
-  {
-    "namespace": "openshift-sdn"
-  },
-  {
-    "namespace": "openshift-service-ca"
-  },
-  {
-    "namespace": "openshift-service-ca-operator"
-  },
-  {
-    "namespace": "openshift-service-catalog-removed"
-  },
-  {
-    "namespace": "openshift-user-workload-monitoring"
-  },
-  {
-    "namespace": "openshift-vsphere-infra"
-  }
-]
